Extract Status component from GroupLink
